Move WatsonTcp to watsontcp_dotnetcore.Tcp namespace, upgrade protocol to 8
